refactor: 따로 분리된 스타일 파일 하나의 컴포넌트로 합치기
