--- src/components/layout/Layout.style.ts
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-export const Header = styled.header`
-  display: flex;
-  width: 100%;
-`;
-
-export const Main = styled.main`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-`;
