--- src/components/layout/footer/Footer.tsx
@@ -1,3 +1,0 @@
-export default function Footer() {
-  return <div>Footer</div>;
-}
