--- src/components/likes/Likes.style.ts
@@ -1,22 +1,0 @@
-import styled from 'styled-components';
-import { GoHeartFill } from 'react-icons/go';
-
-export const LikesStyle = styled.button<{
-  width: string;
-  height: string;
-}>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: ${({ width }) => width};
-  height: ${({ height }) => height};
-  background-color: ${({ theme }) => theme.color.heartBg};
-  border-radius: 50%;
-  border: none;
-`;
-
-export const HeartIcon = styled(GoHeartFill)<{ liked: boolean }>`
-  width: 60%;
-  height: 60%;
-  color: ${({ liked, theme }) => (liked ? theme.color.heartR : theme.color.heartG)};
-`;
